Moving forecast to its own component
