+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import Weather from './Weather';
+import Weather from './components/Weather';
 
 class App extends Component {
   render() {
