+++ client/src/components/Weather.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import Form from './components/Form';
+import Form from './Form';
+import Forecast from './Forecast';
 
 class Weather extends Component {
     constructor(){
@@ -80,27 +81,18 @@ class Weather extends Component {
                     <h1>5-Day Forecast for {location.name}</h1>
                 </header>
 
-                <Form city_input={city_input} handleCityInput={this.handleCityInput} getWeather={this.getWeather} />
+                <Form city_input={city_input} 
+                handleCityInput={this.handleCityInput} 
+                getWeather={this.getWeather} />
                 
                 <section id="temp-choice">
                     <button id="celsius" name="temp_unit" onClick={this.handleTempUnit}>&deg;C</button>
                     <button id="fahrenheit" name="temp_unit" onClick={this.handleTempUnit}>&deg;F</button>
                 </section>
 
-                {forecast.map( (el, i) =>
-                    <section className="forecast" key={i}>
-                        <p id="day">{new Date(el.date_epoch*1000).toUTCString().slice(0, 16)}</p>
-                        <img src={el.day.condition.icon} alt={el.day.condition.text} id="condition-icon"/>
-                        <p id="condition-text">{el.day.condition.text}</p>
-                        <p id="high-temp">High:{" "}{temp_unit === "fahrenheit" ? el.day.maxtemp_f : el.day.maxtemp_c}</p>
-                        <p id="low-temp">Low:{" "}{temp_unit === "fahrenheit" ?el.day.mintemp_f : el.day.mintemp_c}</p>
-                        <p id="precipitation">Precipitation:{" "} {el.day.totalprecip_in} in.</p>
-                        <p id="humidity">Humidity:{" "}{el.day.avghumidity}%</p>
-                        <p id="uv-index">UV Index:{" "}{el.day.uv}</p>
-                        <p id="sunrise">Sunrise:{" "}{el.astro.sunrise}</p>
-                        <p id="sunset">Sunset:{" "}{el.astro.sunset}</p>
-                    </section>
-                    )}
+                <Forecast forecast={forecast} 
+                temp_unit={temp_unit} />
+                
             </section>
         )
     }
